+++ exercises/module7/7.1-2-3/src/main.jsx
@@ -6,6 +6,7 @@ import App from "components/App/App";
 import HomePage from "components/Pages/HomePage";
 import CinemaPage from "components/Pages/CinemaPage";
 import MovieListPage from "components/Pages/MovieListPage";
+import AddMoviePage from "components/Pages/AddMoviePage";
 
 const router = createBrowserRouter([
   {
@@ -23,6 +24,10 @@ const router = createBrowserRouter([
       {
         path: "movies",
         element: <MovieListPage />,
+      },
+      {
+        path: "movies/add",
+        element: <AddMoviePage />,
       },
     ],
   },
